Split SearchForm.tsx into config/controls/component files
SearchForm.tsx was 687 lines, ~2.1x the next-largest file in its
directory (useSearch.ts at 329). Item 5 of the ongoing code-splitting
refactor series (item 1: PlannerCanvasPreview.tsx removal, PR #419;
item 2: api.ts domain-module split, PR #420; item 3: MyWorkWorkspace.tsx
split, PR #421; item 4: R3FMapFoundation.tsx scene-concern split, PR
#422).

Pure code movement, no logic changes, no exported symbols besides
SearchForm/SearchFormProps changed:
- searchFormConfig.ts: SLIDER_GROUPS, FilterSectionId, ECONOMY_OPTIONS,
  and the pure filter-summary helpers (distanceSummary,
  rangeGroupSummary, systemProfileSummary, sortLabel).
- SearchFormControls.tsx: the generic/presentational subcomponents
  (RangeRow, CheckboxRow, WorkspaceSection, FilterWorkspaceHeader,
  QuickPresets).
- SearchForm.tsx (687 -> 432 lines): keeps the orchestrating component,
  its state, and the per-section JSX workspace, now importing from the
  two new files.

Verified: only SearchForm/SearchFormProps are imported outside this
file (FinderView.tsx), and SearchForm.test.tsx exercises it as a black
box, so no other file needed updating. typecheck, lint, knip, the
search-scoped test suite (34 tests incl. SearchForm.test.tsx's 9), and
production build all pass. One unrelated ColonyPlannerWorkspace.test.tsx
failure under full-suite resource contention was confirmed pre-existing
and unrelated by an isolated re-run (17/17 passing).
